CS171 – Let's Make a Deal
This was the fourth homework assignment that I created for my CS 171 class.  In this program, a statistical theory from the famous game show Let's Make a Deal is proven.  The game has three doors with unknown items behind it, one being a coveted item and the other two being unsatisfactory items.  Once the player chooses a door, the host reveals what was behind one of the losing doors, making sure not to reveal the prize or what is behind the player's door.  Then, the host gives the player a chance to either switch their door or keep their door.  Most people may believe that winning the prize is always has a 33% chance of happening.  This program does 10,000 test runs of when the player keeps their door and switches their door and disproves this statistical assumption.