CS171 – BAC Calculator
This was the third homework assignment I created for my CS 171 class.  In this program, the BAC is calculated for the user for up to 10 drinks.  After the user supplies their weight, sex, and duration since last drink, the program will display the user's BAC for one drink up to 10 drinks.  Also, a message will be displayed alongside the BAC to let the user know how serious their BAC level is.