CS171 – Number Conversions
This was the first homework assignment I created for my CS 171 class.  There are three programs included in this file.  The first program calculates the tuition price at a college for five years as a yearly interest is added on for each year.  The second program converts fluid ounces into different units of fluids.  The third program (extra credit) converts the time in Philadelphia in military time to the time in six cities, all of each are in different times zones.